requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
flask-cors>=4.0.0
scapy>=2.5.0
python-nmap>=1.6.0
netifaces>=0.11.0
netaddr>=0.10.1
pyshark>=0.6.0
SQLAlchemy>=2.0.0
apscheduler>=3.10.4
paramiko>=3.4.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
pandas>=2.1.0
tabulate>=0.9.0
rich>=13.7.0
click>=8.1.7
psutil>=5.9.8
requests>=2.31.0
pytest>=8.0.0
pytest-cov>=4.1.0
